feat: :tada: Initial commit
Initial commit: Add project files and assets  This commit adds the initial set of files and assets for the project. It includes the .gitignore file, CMakeLists.txt, NumberConversions.qmlproject, README.md, and various other files and directories. The commit also includes the implementation files for decimal to binary conversion and the necessary QML components.